backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
pydantic_settings
joblib
scikit-learn
seaborn
matplotlib
openpyxl
google-generativeai
python-dotenv